# test(cache): add performance benchmarks